requirements.txt

selenium
Unidecode
requests
pandas
pillow
Flask
flask-sqlalchemy
flask-login
